Core: include optimization docutils
