+++ include/inviwo/core/datastructures/datatraits.h
@@ -31,6 +31,7 @@
 
 #include <inviwo/core/common/inviwocoredefine.h>
 #include <inviwo/core/util/document.h>
+#include <inviwo/core/util/docutils.h>
 #include <inviwo/core/util/formats.h>
 #include <inviwo/core/util/glmvec.h>
 #include <inviwo/core/util/glmfmt.h>
