+++ include/inviwo/core/properties/minmaxproperty.h
@@ -34,6 +34,7 @@
 #include <inviwo/core/util/glmvec.h>
 #include <inviwo/core/util/exception.h>
 #include <inviwo/core/util/stringconversion.h>
+#include <inviwo/core/util/docutils.h>
 
 #include <algorithm>
 #include <limits>
