+++ modules/brushingandlinking/src/ports/brushingandlinkingports.cpp
@@ -29,14 +29,15 @@
 
 #include <modules/brushingandlinking/ports/brushingandlinkingports.h>
 
-#include <inviwo/core/datastructures/bitset.h>                         // for BitSet
-#include <inviwo/core/io/serialization/deserializer.h>                 // for Deserializer
-#include <inviwo/core/io/serialization/serializer.h>                   // for Serializer
-#include <inviwo/core/network/networkutils.h>                          // for getPredecessors
-#include <inviwo/core/ports/inport.h>                                  // for Inport
-#include <inviwo/core/ports/outport.h>                                 // for Outport
-#include <inviwo/core/ports/port.h>                                    // for Port
-#include <inviwo/core/properties/invalidationlevel.h>                  // for InvalidationLevel
+#include <inviwo/core/datastructures/bitset.h>          // for BitSet
+#include <inviwo/core/io/serialization/deserializer.h>  // for Deserializer
+#include <inviwo/core/io/serialization/serializer.h>    // for Serializer
+#include <inviwo/core/network/networkutils.h>           // for getPredecessors
+#include <inviwo/core/ports/inport.h>                   // for Inport
+#include <inviwo/core/ports/outport.h>                  // for Outport
+#include <inviwo/core/ports/port.h>                     // for Port
+#include <inviwo/core/properties/invalidationlevel.h>   // for InvalidationLevel
+#include <inviwo/core/util/docutils.h>
 #include <inviwo/core/util/document.h>                                 // for Document, TableBui...
 #include <inviwo/core/util/statecoordinator.h>                         // for StateCoordinator
 #include <modules/brushingandlinking/brushingandlinkingmanager.h>      // for BrushingAndLinking...
