+++ modules/python3/src/pythoninport.cpp
@@ -31,6 +31,7 @@
 
 #include <modules/python3/pythonoutport.h>
 #include <inviwo/core/util/document.h>
+#include <inviwo/core/util/docutils.h>
 
 namespace inviwo {
