+++ modules/python3/src/pythonoutport.cpp
@@ -29,6 +29,7 @@
 
 #include <modules/python3/pythonoutport.h>
 #include <inviwo/core/util/document.h>
+#include <inviwo/core/util/docutils.h>
 
 namespace inviwo {
